Image switch to size_t for size and Rect2<int> for clip bounds

+++ NerdFramework++/Image3.cpp
@@ -111,7 +111,7 @@ Image3& Image3::operator=(Image3&& rhs) {
 	rhs.data = nullptr;
 	return *this;
 }
-Image3::Image3(const Image3& rhs, const Rect2<size_t>& clipBounds) :
+Image3::Image3(const Image3& rhs, const Rect2<int>& clipBounds) :
 	_width(clipBounds.width),
 	_height(clipBounds.height),
 	_size(_width * _height * 3),
@@ -124,7 +124,7 @@ Image3::Image3(const Image3& rhs, const Rect2<size_t>& clipBounds) :
 		std::copy(rhs.data + beginOffset, rhs.data + endOffset, data + begin);
 	}
 }
-Image3::Image3(Image3&& rhs, const Rect2<size_t>& clipBounds) :
+Image3::Image3(Image3&& rhs, const Rect2<int>& clipBounds) :
 	_width(clipBounds.width),
 	_height(clipBounds.height),
 	_size(_width * _height * 3),
@@ -151,7 +151,7 @@ int Image3::width() const {
 int Image3::height() const {
 	return this->_height;
 }
-uint32_t Image3::size() const {
+size_t Image3::size() const {
 	return this->_size;
 }
